Inventory Entity

+++ server/src/main/java/com/konecta/internship/Restaurant_POS_System/Inventory/InventoryRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InventoryRepository extends JpaRepository<InventoryEntity,Integer> 
+public interface InventoryRepository extends JpaRepository<InventoryEntity,Long> 
 {
 
 }
